feat: 마크다운 텍스트 컴포넌트로 분리
