+++ FE/src/assets/CustomMarkdownText.jsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import styled from 'styled-components';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+import { StyledRestoreCSS } from '../styles/StyledRestoreCSS';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { dark } from 'react-syntax-highlighter/dist/esm/styles/prism';
+
+export default function CustomMarkdownText({ className, content, children }) {
+    return (
+        <StyledRestoreCSS>
+            <StyledReactMarkdown
+                children={content}
+                remarkPlugins={[remarkGfm]}
+                components={{
+                    code({ node, inline, className, children, ...props }) {
+                        const match = /language-(\w+)/.exec(className || '');
+                        return !inline && match ? (
+                            <SyntaxHighlighter
+                                children={String(children).replace(/\n$/, '')}
+                                style={dark}
+                                language={match[1]}
+                                PreTag="div"
+                                {...props}
+                            />
+                        ) : (
+                            <code className={className} {...props}>
+                                {children}
+                            </code>
+                        );
+                    },
+                }}
+            />
+        </StyledRestoreCSS>
+    );
+}
+
+const StyledReactMarkdown = styled(ReactMarkdown)`
+    padding: 15px;
+`;
